refactor: global -> external.s3 패키지 이동
